picoweb: Allow to override "global" exception handling for actions.
.handle_exc() can be overriden to act on exception occuring in any request
handling (or in picoweb dispatching code). This would be a place to send
a "500 Internal Error" page, except that output might have already started,
so it's not that easy... Anyway, added as it was a clearly missing feature.

For small webapps, explicit exception handling on the level of individual
actions is still the recommended way. (And it's always possible to abstract
repetetive code for that into a decorator.)